requirements.txt
streamlit
numpy
opencv-python-headless
pillow
gdown
tensorflow==2.15.0